파일분석 출력
