--- src/main/java/com/cdg/kakaotalk/AnalysisFile.java
@@ -1,6 +1,0 @@
-package com.cdg.kakaotalk;
-
-public class AnalysisFile {
-
-	
-}
