+++ src/main/java/com/cdg/kakaotalk/Analyzer.java
@@ -1,12 +1,13 @@
 package com.cdg.kakaotalk;
 
 import java.util.ArrayList;
+import java.util.Iterator;
 import java.util.List;
-import java.util.ListIterator;
 
 public class Analyzer {
 
 	private FileInput fileInput;
+	private FileOutput fileOutput = new FileOutput();
 	private List<String> lineList = new ArrayList<String>();
 
 	Time time = new Time();
@@ -21,31 +22,32 @@ public class Analyzer {
 
 		lineList = fileInput.read();
 
-		ListIterator<String> it = lineList.listIterator();
+		Iterator<String> it = lineList.listIterator();
 
 		while (it.hasNext()) {
 
 			String str = (String) it.next();
-			String[] sp = str.split("] ", 3);
 
-			time.add(sp[1]);
-
-			String name = people1.getName();
-			// String name = sp[0].substring(1);
-
-			if (people1.getName().equals(name)) {
-				people1.calculate(sp[2]);
-			} else {
-				people2.calculate(sp[2]);
+			if (str.length() == 0 || str.charAt(0) != '[') {
+				// System.out.println(str);
+				continue;
 			}
 
-		}
+			String[] sp = str.split("] ", 3);
 
+			String name = sp[0].substring(1);
+			time.add(sp[1]);
+			if (people1.getName().equals(name)) {
+				people1.add(sp[2]);
+			} else {
+				people2.add(sp[2]);
+			}
+		}
 	}
 
 	public void print() {
-		time.print();
-		people1.print();
-		people2.print();
+		fileOutput.timePrint(time.getMap());
+		fileOutput.peoplePrint(people1.getName(), people1.getMap());
+		fileOutput.peoplePrint(people2.getName(), people2.getMap());
 	}
 }
